+++ changelog.md
@@ -4,6 +4,7 @@
 
 * remove dependencies on 'base-compat' and 'base-compat-batteries' (#318)
 * re-add MonadFail (ActionT m) instance (#325)
+* raise lower bound on base ( > 4.14 ) to reflect support for GHC >= 8.10 (#325).
 
 ## 0.20 [2023.10.02]
 * Drop support for GHC < 8.10 and modernise the CI pipeline (#300).
@@ -19,6 +20,7 @@ Breaking:
 * (#310) Introduce `unliftio` as a dependency, and base exception handling on `catch`.
 * (#310) Clarify the exception handling mechanism of ActionT, ScottyT. `rescue` changes signature to use proper `Exception` types rather than strings. Remove `ScottyError` typeclass.
 * (#310) All ActionT methods (`text`, `html` etc.) have now a MonadIO constraint on the base monad rather than Monad because the response is constructed in a TVar inside ActionEnv. `rescue` has a MonadUnliftIO constraint. The Alternative instance of ActionT also is based on MonadUnliftIO because `<|>` is implemented in terms of `catch`. `ScottyT` and `ActionT` do not have an exception type parameter anymore.
+* (#310) MonadError e (ActionT m) instance removed
 * (#310) MonadFail (ActionT m) instance is missing by mistake.
 
 ## 0.12.1 [2022.11.17]